// src/views/mentor/entrevistas/entrevistas/Types.ts
export interface ListaEntrevistas {
    id: number;
    titulo: string;
    descripcion: string;
    activo: boolean;
}

export interface PaginatedResponse {
    content: ListaEntrevistas[];
    pageable: {
        pageNumber: number;
        pageSize: number;
        sort: {
            empty: boolean;
            sorted: boolean;
            unsorted: boolean;
        };
        offset: number;
        paged: boolean;
        unpaged: boolean;
    };
    last: boolean;
    totalElements: number;
    totalPages: number;
    size: number;
    number: number;
    sort: {
        empty: boolean;
        sorted: boolean;
        unsorted: boolean;
    };
    first: boolean;
    numberOfElements: number;
    empty: boolean;
}